requirements.txt
pytz==2023.3.post1
mmh3==4.0.1
web3==7.7.0
requests==2.32.3
pycryptodome==3.21.0
websockets==12.0
parsimonious==0.9.0
python-dotenv==1.0.0
pandas==2.2.3
